feat(catgory): category crud

+++ src/modules/category/category.service.ts
@@ -4,6 +4,7 @@ import ApiError from '../errors/ApiError';
 import Category from './category.model';
 import { ICategoryDoc, NewCreatedCategory, UpdateCategoryBody } from './category.interfaces';
 import { IOptions, QueryResult } from '../paginate/paginate';
+import { IUserDoc } from '../user/user.interfaces';
 
 /**
  * Create a category
@@ -26,18 +27,54 @@ export const getCategoryById = async (id: mongoose.Types.ObjectId): Promise<ICat
 
 export const updateCategoryById = async (
   categoryId: mongoose.Types.ObjectId,
+  user: IUserDoc,
   updateBody: UpdateCategoryBody
 ): Promise<ICategoryDoc | null> => {
   const category = await getCategoryById(categoryId);
   if (!category) throw new ApiError(httpStatus.NOT_FOUND, 'Category not found.');
+
+  if (user.role !== 'super_admin' && user.businessId.toString() !== category.businessId.toString())
+    throw new ApiError(httpStatus.BAD_REQUEST, 'You can only edit your own categories.');
+
   Object.assign(category, updateBody);
   await category.save();
   return category;
 };
 
-export const deleteCategoryById = async (categoryId: mongoose.Types.ObjectId): Promise<ICategoryDoc | null> => {
+export const deleteCategoryById = async (
+  categoryId: mongoose.Types.ObjectId,
+  user: IUserDoc
+): Promise<ICategoryDoc | null> => {
   const category = await getCategoryById(categoryId);
   if (!category) throw new ApiError(httpStatus.NOT_FOUND, 'Category not found.');
+
+  if (user.role !== 'super_admin' && user.businessId.toString() !== category.businessId.toString())
+    throw new ApiError(httpStatus.BAD_REQUEST, 'You can only edit your own categories.');
+
   await category.deleteOne();
   return category;
+};
+
+export const getCategories = async (businessId: mongoose.Types.ObjectId | null): Promise<ICategoryDoc[]> => {
+  let matchQuery: any = {};
+
+  if (businessId) {
+    matchQuery = {
+      businessId,
+    };
+  }
+
+  const categories = await Category.aggregate([
+    {
+      $match: matchQuery,
+    },
+    {
+      $project: {
+        updatedAt: 0,
+        __v: 0,
+      },
+    },
+  ]);
+
+  return categories;
 };
